chore(app): Rename notifications.dart to firebase.dart
Firebase is not just about push notifications.
